+++ src/Controller/HearingController.php
@@ -7,10 +7,13 @@ use App\Entity\CaseEntity;
 use App\Entity\CaseEvent;
 use App\Entity\DigitalPost;
 use App\Entity\DigitalPostAttachment;
+use App\Entity\Document;
 use App\Entity\Hearing;
+use App\Entity\HearingBriefing;
 use App\Entity\HearingPost;
 use App\Entity\HearingPostRequest;
 use App\Entity\HearingPostResponse;
+use App\Entity\HearingRecipient;
 use App\Entity\User;
 use App\Exception\HearingException;
 use App\Form\HearingFinishType;
@@ -38,7 +41,7 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class HearingController extends AbstractController
 {
-    public function __construct(private EntityManagerInterface $entityManager, private TranslatorInterface $translator)
+    public function __construct(private EntityManagerInterface $entityManager, private DocumentUploader $documentUploader, private CaseDocumentRelationRepository $relationRepository, private TranslatorInterface $translator)
     {
     }
 
@@ -216,27 +219,44 @@ class HearingController extends AbstractController
 
             $hearingPost->setHearing($hearing);
 
-            // Create new file from template
-            $fileName = $mailTemplateHelper->renderMailTemplate($hearingPost->getTemplate(), $hearingPost);
+            // Per chosen recipient, create a HearingRecipient and add it to the HearingPostRequest
+            // HearingRecipient is then responsible for containing the rendered document.
+            foreach ($form->get('recipients')->getData() as $recipient) {
+                $hearingRecipient = new HearingRecipient();
+                $hearingRecipient->setRecipient($recipient);
+                $hearingRecipient->setHearingPostRequest($hearingPost);
 
-            // Create document
-            $document = $documentUploader->createDocumentFromPath($fileName, $hearingPost->getTitle(), 'Hearing');
+                // Create new file from template
+                $fileName = $mailTemplateHelper->renderMailTemplate($hearingPost->getTemplate(), $hearingRecipient);
 
-            $hearingPost->setDocument($document);
+                // Create document
+                $document = $documentUploader->createDocumentFromPath($fileName, $hearingPost->getTitle(), 'Hearing');
 
-            // Create case document relation
-            $relation = new CaseDocumentRelation();
-            $relation->setCase($case);
-            $relation->setDocument($document);
+                $hearingRecipient->setDocument($document);
+
+                // Create case document relation
+                $relation = new CaseDocumentRelation();
+                $relation->setCase($case);
+                $relation->setDocument($document);
+
+                $this->entityManager->persist($relation);
+                $this->entityManager->persist($document);
+                $this->entityManager->persist($hearingRecipient);
+
+                $hearingPost->addHearingRecipient($hearingRecipient);
+            }
 
             $hearing->setHasNewHearingPost(false);
-            $this->entityManager->persist($relation);
-            $this->entityManager->persist($document);
             $this->entityManager->persist($hearingPost);
             $this->entityManager->flush();
+
             $this->addFlash('success', new TranslatableMessage('Hearing post request created', [], 'case'));
 
-            return $this->redirectToRoute('case_hearing_index', ['id' => $case->getId(), 'hearing' => $hearing->getId()]);
+            if ($hearingPost->shouldSendBriefing()) {
+                return $this->redirectToRoute('case_hearing_briefing_create', ['case' => $case->getId(), 'hearing' => $hearing->getId(), 'hearingPost' => $hearingPost->getId()]);
+            } else {
+                return $this->redirectToRoute('case_hearing_index', ['id' => $case->getId(), 'hearing' => $hearing->getId()]);
+            }
         }
 
         return $this->render('case/hearing/post_create.html.twig', [
@@ -321,14 +341,18 @@ class HearingController extends AbstractController
 
         $availableParties = $partyHelper->getRelevantPartiesForHearingPostByCase($case);
         $mailTemplates = $mailTemplateHelper->getTemplates('hearing');
+        $informMailTemplates = $mailTemplateHelper->getTemplates('briefing');
 
         $caseDocuments = $documentRepository->getAvailableDocumentsForCase($case);
 
         $form = $this->createForm(HearingPostRequestType::class, $hearingPost, [
             'case_parties' => $availableParties,
-            'mail_template_choices' => $mailTemplates,
             'available_case_documents' => $caseDocuments,
+            'preselects' => array_map(static fn (HearingRecipient $hearingRecipient) => $hearingRecipient->getRecipient(), $hearingPost->getHearingRecipients()->toArray()),
+            'mail_template_choices' => $mailTemplates,
         ]);
+
+        $briefing = $hearingPost->getBriefing();
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -342,23 +366,65 @@ class HearingController extends AbstractController
 
             $hearingPost->setCustomData($customData);
 
-            // Create new file
-            $fileName = $mailTemplateHelper->renderMailTemplate($hearingPost->getTemplate(), $hearingPost);
+            // We do it the slow way by simply just remaking the documents
+            foreach ($hearingPost->getHearingRecipients() as $hearingRecipient) {
+                $document = $hearingRecipient->getDocument();
 
-            // For now, we just overwrite completely
-            $documentUploader->replaceFileContent($hearingPost->getDocument(), $fileName);
+                $this->removeDocumentFromCase($case, $document);
 
-            // Update Document
-            /** @var User $user */
-            $user = $this->getUser();
-            $hearingPost->getDocument()->setDocumentName($hearingPost->getTitle());
-            $hearingPost->getDocument()->setUploadedBy($user);
-            $hearingPost->getDocument()->setUploadedAt(new DateTime('now'));
+                $hearingPost->removeHearingRecipient($hearingRecipient);
+
+                $this->entityManager->remove($hearingRecipient);
+                $this->entityManager->flush();
+            }
+
+            // Per chosen recipient, create a HearingRecipient and add it to the HearingPostRequest
+            // HearingRecipient is then responsible for containing the rendered document.
+            foreach ($form->get('recipients')->getData() as $recipient) {
+                $hearingRecipient = new HearingRecipient();
+                $hearingRecipient->setRecipient($recipient);
+                $hearingRecipient->setHearingPostRequest($hearingPost);
+
+                // Create new file from template
+                $fileName = $mailTemplateHelper->renderMailTemplate($hearingPost->getTemplate(), $hearingRecipient);
+
+                // Create document
+                $document = $documentUploader->createDocumentFromPath($fileName, $hearingPost->getTitle(), 'Hearing');
+
+                $hearingRecipient->setDocument($document);
+
+                // Create case document relation
+                $relation = new CaseDocumentRelation();
+                $relation->setCase($case);
+                $relation->setDocument($document);
+
+                $this->entityManager->persist($relation);
+                $this->entityManager->persist($document);
+                $this->entityManager->persist($hearingRecipient);
+
+                $hearingPost->addHearingRecipient($hearingRecipient);
+            }
 
             $this->entityManager->flush();
             $this->addFlash('success', new TranslatableMessage('Hearing post request updated', [], 'case'));
 
-            return $this->redirectToRoute('case_hearing_post_show', ['case' => $case->getId(), 'hearingPost' => $hearingPost->getId()]);
+            if ($hearingPost->shouldSendBriefing()) {
+                if (null === $briefing) {
+                    return $this->redirectToRoute('case_hearing_briefing_create', ['case' => $case->getId(), 'hearing' => $hearingPost->getHearing()->getId(), 'hearingPost' => $hearingPost->getId()]);
+                } else {
+                    return $this->redirectToRoute('case_hearing_briefing_edit', ['case' => $case->getId(), 'hearing' => $hearingPost->getHearing()->getId(), 'hearingPost' => $hearingPost->getId(), 'briefing' => $briefing->getId()]);
+                }
+            } else {
+                if (null != $briefing) {
+                    // Necessary due to cascade
+                    $hearingPost->setBriefing(null);
+                    $briefing->setHearingPostRequest(null);
+
+                    $this->removeBriefing($briefing);
+                }
+
+                return $this->redirectToRoute('case_hearing_post_show', ['case' => $case->getId(), 'hearingPost' => $hearingPost->getId()]);
+            }
         }
 
         return $this->render('case/hearing/post_edit.html.twig', [
@@ -371,7 +437,7 @@ class HearingController extends AbstractController
     /**
      * @Route("/{case}/hearing/{hearingPost}/approve", name="case_hearing_post_approve", methods={"POST"})
      */
-    public function hearingPostApprove(CaseEntity $case, HearingPostResponse $hearingPost, MailTemplateHelper $mailTemplateHelper, DocumentUploader $documentUploader, DigitalPostHelper $digitalPostHelper, CaseEventHelper $caseEventHelper): Response
+    public function hearingPostApprove(CaseEntity $case, HearingPostResponse $hearingPost, CaseEventHelper $caseEventHelper): Response
     {
         $this->denyAccessUnlessGranted('edit', $case);
 
@@ -421,19 +487,57 @@ class HearingController extends AbstractController
             $digitalPostAttachments[] = $digitalPostAttachment;
         }
 
-        // Handle recipients
-        $digitalPostRecipients = [];
+        foreach ($hearingPost->getHearingRecipients() as $hearingRecipient) {
+            $digitalPostRecipient = (new DigitalPost\Recipient())
+                ->setName($hearingRecipient->getRecipient()->getName())
+                ->setIdentifierType($hearingRecipient->getRecipient()->getIdentification()->getType())
+                ->setIdentifier($hearingRecipient->getRecipient()->getIdentification()->getIdentifier())
+                ->setAddress($hearingRecipient->getRecipient()->getAddress())
+            ;
 
-        $digitalPostRecipients[] = (new DigitalPost\Recipient())
-            ->setName($hearingPost->getRecipient()->getName())
-            ->setIdentifierType($hearingPost->getRecipient()->getIdentification()->getType())
-            ->setIdentifier($hearingPost->getRecipient()->getIdentification()->getIdentifier())
-            ->setAddress($hearingPost->getRecipient()->getAddress())
-        ;
+            $clonedAttachments = array_map(static fn (DigitalPostAttachment $attachment) => (new DigitalPostAttachment())->setDigitalPost($attachment->getDigitalPost())->setDocument($attachment->getDocument()), $digitalPostAttachments);
 
-        $digitalPost = $digitalPostHelper->createDigitalPost($hearingPost->getDocument(), $hearingPost->getTitle(), get_class($case), $case->getId(), $digitalPostAttachments, $digitalPostRecipients);
+            $digitalPost = $digitalPostHelper->createDigitalPost($hearingRecipient->getDocument(), $hearingPost->getTitle(), get_class($case), $case->getId(), $clonedAttachments, [$digitalPostRecipient]);
 
-        $caseEventHelper->createDigitalPostCaseEvent($case, $digitalPost, [$hearingPost->getRecipient()]);
+            $caseEventHelper->createDigitalPostCaseEvent($case, $digitalPost, $hearingPost->getTitle(), [$hearingRecipient->getRecipient()]);
+        }
+
+        // Now handle potential briefings.
+        if ($hearingPost->shouldSendBriefing()) {
+            $briefing = $hearingPost->getBriefing();
+            if (!$briefing) {
+                throw new HearingException('Attempting to send empty briefing');
+            }
+
+            $briefingRecipients = $briefing->getHearingBriefingRecipients();
+            if ($briefingRecipients->isEmpty()) {
+                throw new HearingException('Could not find any recipients for briefing');
+            }
+
+            $briefingAttachments = $briefing->getAttachments();
+            $digitalPostBriefingAttachments = [];
+
+            foreach ($briefingAttachments as $briefingAttachment) {
+                $digitalPostAttachment = new DigitalPostAttachment();
+                $digitalPostAttachment->setDocument($briefingAttachment);
+                $digitalPostBriefingAttachments[] = $digitalPostAttachment;
+            }
+
+            foreach ($briefingRecipients as $briefingRecipient) {
+                $digitalPostBriefingRecipient = (new DigitalPost\Recipient())
+                    ->setName($briefingRecipient->getRecipient()->getName())
+                    ->setIdentifierType($briefingRecipient->getRecipient()->getIdentification()->getType())
+                    ->setIdentifier($briefingRecipient->getRecipient()->getIdentification()->getIdentifier())
+                    ->setAddress($briefingRecipient->getRecipient()->getAddress())
+                ;
+
+                $clonedBriefingAttachments = array_map(static fn (DigitalPostAttachment $attachment) => (new DigitalPostAttachment())->setDigitalPost($attachment->getDigitalPost())->setDocument($attachment->getDocument()), $digitalPostBriefingAttachments);
+
+                $digitalPost = $digitalPostHelper->createDigitalPost($briefingRecipient->getDocument(), $briefing->getTitle(), get_class($case), $case->getId(), $clonedBriefingAttachments, [$digitalPostBriefingRecipient]);
+
+                $caseEventHelper->createDigitalPostCaseEvent($case, $digitalPost, $briefing->getTitle(), [$briefingRecipient->getRecipient()]);
+            }
+        }
 
         $today = new DateTime('today');
 
@@ -482,34 +586,41 @@ class HearingController extends AbstractController
     /**
      * @Route("/{case}/hearing/{hearingPost}/delete", name="case_hearing_post_delete")
      */
-    public function hearingPostDelete(CaseEntity $case, HearingPost $hearingPost, DocumentUploader $documentUploader, CaseDocumentRelationRepository $relationRepository, Request $request): Response
+    public function hearingPostDelete(CaseEntity $case, HearingPost $hearingPost, Request $request): Response
     {
         $this->denyAccessUnlessGranted('edit', $case);
 
         // Check that CSRF token is valid
         if ($this->isCsrfTokenValid('delete'.$hearingPost->getId(), $request->request->get('_token'))) {
-            // Remove relation between case and document
-            $document = $hearingPost->getDocument();
-            $relation = $relationRepository->findOneBy(['case' => $case, 'document' => $document]);
+            if ($hearingPost instanceof HearingPostRequest) {
+                foreach ($hearingPost->getHearingRecipients() as $hearingRecipient) {
+                    $document = $hearingRecipient->getDocument();
 
-            if (null !== $relation) {
-                $this->entityManager->remove($relation);
-                $this->entityManager->flush();
+                    $this->removeDocumentFromCase($case, $document);
+
+                    $hearingPost->removeHearingRecipient($hearingRecipient);
+                    $this->entityManager->remove($hearingRecipient);
+                }
+
+                if ($briefing = $hearingPost->getBriefing()) {
+                    $this->removeBriefing($briefing);
+                }
+            } elseif ($hearingPost instanceof HearingPostResponse) {
+                // Remove relation between case and document
+                $document = $hearingPost->getDocument();
+
+                $this->removeDocumentFromCase($case, $document);
+
+                $hearingPost->setDocument(null);
+            } else {
+                throw new HearingException(sprintf('Unhandled HearingPost of type %s detected during deletion.', get_class($hearingPost)));
             }
-
-            // Remove file
-            $documentUploader->deleteDocumentFile($document);
-
-            // Remove reference between post and document
-            $hearingPost->setDocument(null);
-            $this->entityManager->flush();
 
             // Remove new hearing post alert
             $hearingPost->getHearing()->setHasNewHearingPost(false);
 
             // Remove hearing post and document
             $this->entityManager->remove($hearingPost);
-            $this->entityManager->remove($document);
             $this->entityManager->flush();
 
             $message = match (get_class($hearingPost)) {
@@ -521,5 +632,33 @@ class HearingController extends AbstractController
         }
 
         return $this->redirectToRoute('case_hearing_index', ['id' => $case->getId()]);
+    }
+
+    public function removeDocumentFromCase(CaseEntity $case, Document $document)
+    {
+        $relation = $this->relationRepository->findOneBy(['case' => $case, 'document' => $document]);
+
+        if (null !== $relation) {
+            $this->entityManager->remove($relation);
+            $this->entityManager->flush();
+        }
+
+        // Remove file and document
+        $this->documentUploader->deleteDocumentFile($document);
+        $this->entityManager->remove($document);
+    }
+
+    private function removeBriefing(HearingBriefing $briefing)
+    {
+        $briefingRecipients = $briefing->getHearingBriefingRecipients();
+        foreach ($briefingRecipients as $briefingRecipient) {
+            $document = $briefingRecipient->getDocument();
+
+            $this->removeDocumentFromCase($briefing->getHearingPostRequest()->getHearing()->getCaseEntity(), $document);
+            $this->entityManager->remove($briefingRecipient);
+        }
+
+        $this->entityManager->remove($briefing);
+        $this->entityManager->flush();
     }
 }